+++ src/app/business/app-auth/app-user/app-user.component.ts
@@ -28,21 +28,20 @@ export class AppUserComponent implements OnInit {
   }
 
   ngOnInit() {
-    this.headers = this.http.getHeader('users');
-
-    this.getHeartData(this.url);
-    this.http.isHavePerm('app-user-del').then(v => {
-      this.deleteBtn = v;
-      this.deleteAllBtn = v;
-    });
-
-    this.http.isHavePerm('app-user-add').then(v => {
-      this.addBtn = v;
-    });
-    this.http.isHavePerm('app-user-edit').then(v => {
-      this.editBtn = v;
-    });
-
+    if(this.http.hasToken()){
+        this.headers = this.http.getHeader('users');
+        this.getHeartData(this.url);
+        this.http.isHavePerm('app-user-del').then(v => {
+            this.deleteBtn = v;
+            this.deleteAllBtn = v;
+        });
+        this.http.isHavePerm('app-user-add').then(v => {
+            this.addBtn = v;
+        });
+        this.http.isHavePerm('app-user-edit').then(v => {
+            this.editBtn = v;
+        });
+    }
   }
 
   headers: Array<cell> = [];
